requirements.txt
google-generativeai
google-cloud-storage
moviepy